+++ Nodify/Connections/Connection.cs
@@ -42,7 +42,7 @@ namespace Nodify
                 var to = InterpolateCubicBezier(p0, p1, p2, p3, t);
                 var direction = GetBezierTangent(p0, p1, p2, p3, t);
 
-                base.DrawDirectionalArrowheadGeometry(context, direction, to);
+                DrawDirectionalArrowheadGeometry(context, direction, to);
             }
         }