+++ Nodify/Editor/NodifyEditor.Selecting.cs
@@ -292,6 +292,8 @@ namespace Nodify
             selected.Clear();
             selected.Add(container.DataContext);
             EndUpdateSelectedItems();
+
+            UnselectAllConnections();
         }
 
         /// <summary>
@@ -360,7 +362,6 @@ namespace Nodify
                 return;
             }
 
-            UnselectAllConnections();
             SelectedArea = _selection.Start(ItemContainers, location, type, EnableRealtimeSelection);
             IsSelecting = true;
         }
@@ -396,9 +397,14 @@ namespace Nodify
                 return;
             }
 
+            if (_selection.Type == SelectionType.Replace)
+            {
+                UnselectAllConnections();
+            }
+
             SelectedArea = _selection.End();
-            ApplyPreviewingSelection();
             IsSelecting = false;
+            ApplyPreviewingSelection();
         }
 
         /// <summary>
@@ -416,7 +422,7 @@ namespace Nodify
 
             if (IsSelecting)
             {
-                ClearPreviewingSelection();
+                _selection.Cancel();
                 IsSelecting = false;
             }
         }
@@ -441,16 +447,6 @@ namespace Nodify
                 container.IsPreviewingSelection = null;
             }
             EndUpdateSelectedItems();
-        }
-
-        private void ClearPreviewingSelection()
-        {
-            ItemCollection items = Items;
-            for (var i = 0; i < items.Count; i++)
-            {
-                var container = (ItemContainer)ItemContainerGenerator.ContainerFromIndex(i);
-                container.IsPreviewingSelection = null;
-            }
         }
 
         #endregion