+++ Nodify/Editor/NodifyEditor.cs
@@ -573,6 +573,8 @@ namespace Nodify
             SetValue(ViewportTransformPropertyKey, transform);
 
             InputProcessor.AddSharedHandlers(this);
+
+            Unloaded += OnEditorUnloaded;
         }
 
         /// <inheritdoc />
@@ -584,6 +586,11 @@ namespace Nodify
             ConnectionsHost = GetTemplateChild(ElementConnectionsHost) as UIElement ?? throw new InvalidOperationException($"{ElementConnectionsHost} is missing or is not of type UIElement.");
 
             OnDisableAutoPanningChanged(DisableAutoPanning);
+        }
+
+        private void OnEditorUnloaded(object sender, RoutedEventArgs e)
+        {
+            OnDisableAutoPanningChanged(true);
         }
 
         /// <inheritdoc />